+++ changelog.md
@@ -1,6 +1,7 @@
 ##### Version 1.21.1-12.3.1-pre
 - Add an item model for the resonanz observer (BluSunrize)
 - Re-add Jade compatability (BluSunrize)
+- Change item routers to no longer allow items out an unfiltered side *if* a filtered side for that item exists (BluSunrize)
 - Fix desync of item router buttons on multiplayer servers (BluSunrize)
 - Fix resonanz observer not rendering the completed structure in the manual (BluSunrize)